// src/app/shared/types/api-types/flight-dto.interface.ts
import { EquipmentDTO } from './equipment-dto.interface';
import { FlightCrewDTO } from './flight-crew-dto.interface';
import { FlightDTO } from './flight-dto.interface';
import { FlightFuelDTO } from './flight-fuel-dto.interface';
import { FlightLoadDTO } from './flight-load-dto.interface';
import { FlightMeasureDTO } from './flight-measure-dto.interface';
import { FlightTimeDTO } from './flight-time-dto.interface';
import { TimezoneDTO } from './timezone-dto.interface';

export interface FlightDTO {
  id?: number;
  number?: string;
  missionCode?: string;
  status?: string;
  type?: string;
  flightRuleId?: number;
  flightRuleAppId?: number;
  missionArea?: string;
  startDate?: Date;
  endDate?: Date;
  startBase?: string;
  endBase?: string;
  flightTime?: string;
  operatingTime?: string;
  landing?: number;
  customerCode?: string;
  categoryName?: string;
  duration?: string;
  durationIFR?: string;
  timezone?: TimezoneDTO;
  nbCycles?: string;
  occurrence?: number;
  nightMission?: boolean;
  shipmentNumber?: number;
  missionNumber?: string;
  missionComment?: string;
  missionStartDate?: Date;
  missionEndDate?: Date;
  missionDuration?: string;
  missionCycleNumber?: string;
  missionPeriodicityType?: string;
  missionPeriodicityLimitDate?: Date;
  missionDescription?: string;
  missionDetails?: string;
  missionOperationCode?: string;
  operationalMissionCode?: string;
  newMission?: boolean;
  previousFlight?: FlightDTO;
  nextFlight?: FlightDTO;
  aircraft?: EquipmentDTO;
  engineStartDate?: Date;
  engineStopDate?: Date;
  usedFuel?: number;
  engine1AssetCode?: string;
  engine2AssetCode?: string;
  engine3AssetCode?: string;
  engine4AssetCode?: string;
  engine1AssetText?: string;
  engine2AssetText?: string;
  engine3AssetText?: string;
  engine4AssetText?: string;
  engine1CycleNumber?: string;
  engine2CycleNumber?: string;
  engine3CycleNumber?: string;
  engine4CycleNumber?: string;
  engine1HourNumber?: string;
  engine2HourNumber?: string;
  engine3HourNumber?: string;
  engine4HourNumber?: string;
  userFlightNumber?: string;
  flightType?: string;
  flightLabel?: string;
  flightDepartureLocation?: string;
  flightArrivalLocation?: string;
  flightDetails?: string;
  flightRequestedOperationalConfig?: string;
  flightTimeList?: FlightTimeDTO[];
  flightMeasuresList?: FlightMeasureDTO[];
  flightMeasuresUpdatable?: boolean;
  flightCrewList?: FlightCrewDTO[];
  flightPilot?: string;
  flightCopilot?: string;
  flightCrew?: string;
  flightFuelList?: FlightFuelDTO[];
  flightFuelLoadedList?: FlightFuelDTO[];
  plannedPax?: number;
  aboardPax?: number;
  deplannedPax?: number;
  loadWeightIomId?: number;
  loadInternalLoad?: number;
  loadPaxLoad?: number;
  loadFuel?: number;
  loadExternalLoad?: number;
  loadAcEmptyWeight?: number;
  loadTotalToWeight?: number;
  flightLoadList?: FlightLoadDTO[];
  environment?: string;
  comment?: string;
}